fix(Package): allow use of `npm link` with dgeni
Use of `instanceof` and `npm link` together often breaks node applications,
since you might end up with multiple versions of the same library in your
node_modules tree. (NodeJS isn't able to look past links to ensure that only
one copy is installed.) This means that `instanceof` might fail for what
appears to be an instance of the correct class (e.g. in this project `Package`).

This commit fixes this in to places.

* First it fixes it for `Package` types by using duck-type checking instead
of `instanceof` (we are only interested in a few properties of Package anyway).
* Second, since the `di` library suffers from the same problem, and is not
going to get fixed, we workaround it by passing plain objects instead of
instances of `Module` to the `di` library.