refactor(*): rewrite in TypeScript
The entire code base for the dgeni tool (not dgeni-packages yet) has been
rewritten in TypeScript.

Clients that use TypeScript in their own projects should be able to get
type information for the Dgeni API.

There should be no breaking changes for clients but please do report any
anomaly that pops up in the issue tracker.